Change structure - we now have option to support different build. Not only cmake + ninja, but any other build system. We will need it for ICU
