+++ Sources/SwiftBuilder/Steps/ConfigureRepoStep.swift
@@ -3,82 +3,12 @@ import Foundation
 import Logging
 import Shell
 
-protocol BuildableItemDependency {
-    func cmakeDepDirCaheEntry(depName: String, config: BuildConfig) -> [String]
-}
-
-extension BuildableItemDependency where Self: BuildableItem {
-    func cmakeDepDirCaheEntry(depName: String, config: BuildConfig) -> [String] {
-        let depBuildUrl = config.buildLocation(for: self)
-        let res = depName + "_DIR=\"\(depBuildUrl.path)/cmake/modules\""
-        return [res]
-    }
-}
-
-struct BuildableItemRepo {
-    let checkoutable: Checkoutable
-    let patchFileName: String
-}
-
-protocol BuildableItem {
-    // Will be used as folder or file name part, where needed
-    var name: String { get }
-
-    var buildSubfolder: String? { get }
-
-    var dependencies: [String: BuildableItemDependency] { get }
-
-    var targets: [String] { get }
-
-    var underlyingRepo: BuildableItemRepo? { get }
-
-    func sourceLocation(using buildConfig: BuildConfig) -> URL
-
-    func cmakeCacheEntries(config: BuildConfig) -> [String]
-}
-
-extension BuildableItem {
-    var buildSubfolder: String? { nil }
-
-    var dependencies: [String: BuildableItemDependency] { [:] }
-
-    // Most of repos has 1 default target
-    var targets: [String] { [] }
-
-    var patchFileName: String { name }
-
-    func cmakeCacheEntries(config: BuildConfig) -> [String] { [] }
-}
-
-extension BuildableItem where Self: Checkoutable {
-    var name: String { repoName }
-
-    func sourceLocation(using buildConfig: BuildConfig) -> URL {
-        var resUrl = buildConfig.location(for: self)
-        if let buildSubfolder {
-            resUrl = resUrl.appendingPathComponent(buildSubfolder, isDirectory: true)
-        }
-        return resUrl
-    }
-
-    var underlyingRepo: BuildableItemRepo? {
-        BuildableItemRepo(checkoutable: self,
-                          patchFileName: repoName + ".patch")
-    }
-}
-
-extension BuildConfig {
-    func buildLocation(for repo: BuildableItem) -> URL {
-        return buildsRootFolder.appendingPathComponent(repo.name, isDirectory: true)
-    }
-}
-
 final class ConfigureRepoStep: BuildStep {
     var stepName: String {
         "configure-" + buildableItem.name
     }
 
-    init(buildableItem: BuildableItem) {
+    init(buildableItem: NinjaBuildableItem) {
         self.buildableItem = buildableItem
     }
 
@@ -137,7 +67,7 @@ final class ConfigureRepoStep: BuildStep {
 
     // MARK: Private
 
-    private let buildableItem: BuildableItem
+    private let buildableItem: NinjaBuildableItem
     private var fileManager: FileManager { FileManager.default }
     private let terminal = Terminal()
     private let defaultRevisionsMap = DefaultRevisionsMap()
