+++ Sources/SwiftBuilder/SwiftBuildCommand.swift
@@ -136,11 +136,7 @@ final class SwiftBuildCommand: AsyncParsableCommand {
     private static let steps: [BuildStep] = {
         let checkoutStep: BuildStep = CheckoutStep(checkoutables: Repos.checkoutOrder)
 
-        let buildSteps: [BuildStep] = Repos.buildOrder.flatMap { repo -> [BuildStep] in
-            let configure = ConfigureRepoStep(buildableItem: repo)
-            let build = NinjaBuildStep(buildableRepo: repo)
-            return [configure, build]
-        }
+        let buildSteps: [BuildStep] = Builds.buildOrder.flatMap { $0.buildSteps() }
 
         return [checkoutStep] + buildSteps
     }()
